refactor: streamline main.go  & better error mgmt.
